Workinprog
Changed the color values to the proper scale. The color value range is > or = 127. 127 being the max for a color value. I determined the new color values with simple %'s . what i mean by this is that since the color values we where using before was "  color value range is > or = 256" and that was wrong i tried to think of a way to change the color values to the proper numbers without cahngeing what we wanted the color to be. What i figured would work was since 127, witch is the max for the proper color value, is roughly 50% of the old color value ,witch was max value 256, i would take all the existing color values that are above 127 ( and all the numbers that are in that rgb bracket with it ) and cut them in half. So now the color values should be right and the color its self should not of changed. Of course since any part of the code that uses those numbers is not working yet , I have no way to prove this but I'm pretty sure I'm right about this.