Workinprog
Since there was already an else statement declaring what needs to happen to have the DisabledRobot i found the having a bit value for it on top of the else statement to be unnecessary. So i got rid of the Serial.println(retValue) lines for it and the if statement that goes with it. It still works fine.