Update LED_Lights.ino
Cleaned up code.

Placed the retValue checks in if/else block so to improve efficiency. Placed checks in numerical order so it's easier to read.

There was a bug in that if retValue didn't add up to one of the many values checked by the if statements... it would still call strip.show(), but it would likely leave the strip displaying colors that are no longer applicable (whatever was shown last).